itoa done

+++ ft_atoi.c
@@ -6,14 +6,14 @@
 /*   By: tmarts <tmarts@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 22:36:55 by tmarts            #+#    #+#             */
-/*   Updated: 2022/10/24 20:17:34 by tmarts           ###   ########.fr       */
+/*   Updated: 2022/10/26 18:47:14 by tmarts           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_atoi(const char *str)
 {
-	long int	intvalue;
-	int			neg_pos;
+	long long int	intvalue;
+	int				neg_pos;
 
 	intvalue = 0;
 	neg_pos = 1;
